+++ docs/src/config.ts
@@ -9,17 +9,20 @@ export const SIDEBAR = {
 
     { text: 'Basics', header: true },
     { text: 'Project Structure', link: 'core-concepts/project-structure' },
-    { text: 'Components', link: 'core-concepts/astro-components' },
+    { text: 'Component Syntax', link: 'core-concepts/astro-components' },
     { text: 'Pages', link: 'core-concepts/astro-pages' },
     { text: 'Layouts', link: 'core-concepts/layouts' },
-    { text: 'Collections', link: 'core-concepts/collections' },
+    { text: 'Routing', link: 'core-concepts/routing' },
     { text: 'Partial Hydration', link: 'core-concepts/component-hydration' },
 
     { text: 'Guides', header: true },
     { text: 'Styling & CSS', link: 'guides/styling' },
-    { text: 'Data Fetching', link: 'guides/data-fetching' },
     { text: 'Markdown', link: 'guides/markdown-content' },
+    { text: 'Data Fetching', link: 'guides/data-fetching' },
+    { text: 'Pagination', link: 'guides/pagination' },
+    { text: 'RSS', link: 'guides/rss' },
     { text: 'Supported Imports', link: 'guides/imports' },
+    { text: 'Aliases', link: 'guides/aliases' },
     { text: 'Deploy a Website', link: 'guides/deploy' },
     { text: 'Publish a Component', link: 'guides/publish-to-npm' },
 
@@ -33,6 +36,10 @@ export const SIDEBAR = {
     },
     { text: 'Renderer Reference', link: 'reference/renderer-reference' },
   ],
+  de: [
+    { text: 'Willkommen', header: true },
+    { text: 'Einführung', link: 'de/getting-started' },
+  ],
   nl: [
     { text: 'Welkom', header: true },
     { text: 'Beginnen', link: 'nl/getting-started' },
@@ -43,6 +50,10 @@ export const SIDEBAR = {
     { text: 'Pika-aloitus', link: 'fi/quick-start' },
     { text: 'Asennus', link: 'fi/installation' },
   ],
+  es: [
+    { text: 'Configuración', header: true },
+    { text: 'Empezando', link: 'es/getting-started' },
+  ],
   'zh-CN': [
     { text: '起步', header: true },
     { text: '入门指南', link: 'zh-CN/getting-started' },
@@ -51,8 +62,21 @@ export const SIDEBAR = {
     { text: '模板样例', link: 'zh-CN/examples' },
   ],
   'zh-TW': [
-    { text: '起步', header: true },
-    { text: '入門指南', link: 'zh-TW/getting-started' },
+    { text: '設定', header: true },
+    { text: '新手上路', link: 'zh-TW/getting-started' },
+    { text: '快速開始', link: 'zh-TW/quick-start' },
+    { text: '安裝', link: 'zh-TW/installation' },
+    { text: '範例', link: 'zh-TW/examples' },
+  ],
+  bg: [
+    { text: 'Главни', header: true },
+    { text: 'Започваме!', link: 'bg/getting-started' },
+  ],
+  fr: [
+    { text: 'Bienvenue', header: true },
+    { text: 'Bien démarrer', link: 'fr/getting-started' },
+    { text: 'Démarrage rapide', link: 'fr/quick-start' },
+    { text: 'Installation', link: 'fr/installation' },
   ],
 };