+++ packages/astro/test/astro-search.test.js
@@ -12,30 +12,18 @@ Search('Finds the root page', async ({ runtime }) => {
 });
 
 Search('Matches pathname to filename', async ({ runtime }) => {
-  const result = await runtime.load('/news');
-  assert.equal(result.statusCode, 200);
+  assert.equal((await runtime.load('/news')).statusCode, 200);
+  assert.equal((await runtime.load('/news/')).statusCode, 200);
 });
 
-Search('A URL with a trailing slash can match a folder with an index.astro', async ({ runtime }) => {
-  const result = await runtime.load('/nested-astro/');
-  assert.equal(result.statusCode, 200);
+Search('Matches pathname to a nested index.astro file', async ({ runtime }) => {
+  assert.equal((await runtime.load('/nested-astro')).statusCode, 200);
+  assert.equal((await runtime.load('/nested-astro/')).statusCode, 200);
 });
 
-Search('A URL with a trailing slash can match a folder with an index.md', async ({ runtime }) => {
-  const result = await runtime.load('/nested-md/');
-  assert.equal(result.statusCode, 200);
-});
-
-Search('A URL without a trailing slash can redirect to a folder with an index.astro', async ({ runtime }) => {
-  const result = await runtime.load('/nested-astro');
-  assert.equal(result.statusCode, 301);
-  assert.equal(result.location, '/nested-astro/');
-});
-
-Search('A URL without a trailing slash can redirect to a folder with an index.md', async ({ runtime }) => {
-  const result = await runtime.load('/nested-md');
-  assert.equal(result.statusCode, 301);
-  assert.equal(result.location, '/nested-md/');
+Search('Matches pathname to a nested index.md file', async ({ runtime }) => {
+  assert.equal((await runtime.load('/nested-md')).statusCode, 200);
+  assert.equal((await runtime.load('/nested-md/')).statusCode, 200);
 });
 
 Search.run();