+++ tools/language-server/src/plugins/typescript/TypeScriptPlugin.ts
@@ -1,21 +1,33 @@
-import { Document, DocumentManager, isInsideFrontmatter } from '../../core/documents';
 import type { ConfigManager } from '../../core/config';
 import type { CompletionsProvider, AppCompletionItem, AppCompletionList } from '../interfaces';
+import type { CancellationToken, Hover, SignatureHelp, SignatureHelpContext } from 'vscode-languageserver';
+import { join as pathJoin, dirname as pathDirname } from 'path';
+import { Document, DocumentManager, isInsideFrontmatter } from '../../core/documents';
+import { SourceFile, ImportDeclaration, Node, SyntaxKind } from 'typescript';
 import { CompletionContext, DefinitionLink, FileChangeType, Position, LocationLink } from 'vscode-languageserver';
 import * as ts from 'typescript';
-import { CompletionsProviderImpl, CompletionEntryWithIdentifer } from './features/CompletionsProvider';
 import { LanguageServiceManager } from './LanguageServiceManager';
 import { SnapshotManager } from './SnapshotManager';
-import { convertToLocationRange, isVirtualFilePath, getScriptKindFromFileName } from './utils';
-import { isNoTextSpanInGeneratedCode, SnapshotFragmentMap } from './features/utils';
+import { convertToLocationRange, isVirtualAstroFilePath, isVirtualFilePath, getScriptKindFromFileName } from './utils';
 import { isNotNullOrUndefined, pathToUrl } from '../../utils';
+import { CompletionsProviderImpl, CompletionEntryWithIdentifer } from './features/CompletionsProvider';
+import { HoverProviderImpl } from './features/HoverProvider';
+import { isNoTextSpanInGeneratedCode, SnapshotFragmentMap } from './features/utils';
+import { SignatureHelpProviderImpl } from './features/SignatureHelpProvider';
+
+type BetterTS = typeof ts & {
+  getTouchingPropertyName(sourceFile: SourceFile, pos: number): Node;
+};
 
 export class TypeScriptPlugin implements CompletionsProvider {
   private readonly docManager: DocumentManager;
   private readonly configManager: ConfigManager;
   private readonly languageServiceManager: LanguageServiceManager;
+  public pluginName = 'TypeScript';
 
   private readonly completionProvider: CompletionsProviderImpl;
+  private readonly hoverProvider: HoverProviderImpl;
+  private readonly signatureHelpProvider: SignatureHelpProviderImpl;
 
   constructor(docManager: DocumentManager, configManager: ConfigManager, workspaceUris: string[]) {
     this.docManager = docManager;
@@ -23,6 +35,12 @@ export class TypeScriptPlugin implements CompletionsProvider {
     this.languageServiceManager = new LanguageServiceManager(docManager, configManager, workspaceUris);
 
     this.completionProvider = new CompletionsProviderImpl(this.languageServiceManager);
+    this.hoverProvider = new HoverProviderImpl(this.languageServiceManager);
+    this.signatureHelpProvider = new SignatureHelpProviderImpl(this.languageServiceManager);
+  }
+
+  async doHover(document: Document, position: Position): Promise<Hover | null> {
+    return this.hoverProvider.doHover(document, position);
   }
 
   async getCompletions(document: Document, position: Position, completionContext?: CompletionContext): Promise<AppCompletionList<CompletionEntryWithIdentifer> | null> {
@@ -46,26 +64,38 @@ export class TypeScriptPlugin implements CompletionsProvider {
     const filePath = tsDoc.filePath;
     const tsFilePath = filePath.endsWith('.ts') ? filePath : filePath + '.ts';
 
-    const defs = lang.getDefinitionAndBoundSpan(tsFilePath, mainFragment.offsetAt(mainFragment.getGeneratedPosition(position)));
+    const fragmentPosition = mainFragment.getGeneratedPosition(position);
+    const fragmentOffset = mainFragment.offsetAt(fragmentPosition);
+
+    let defs = lang.getDefinitionAndBoundSpan(tsFilePath, fragmentOffset);
 
     if (!defs || !defs.definitions) {
       return [];
+    }
+
+    // Resolve all imports if we can
+    if (this.goToDefinitionFoundOnlyAlias(tsFilePath, defs.definitions!)) {
+      let importDef = this.getGoToDefinitionRefsForImportSpecifier(tsFilePath, fragmentOffset, lang);
+      if (importDef) {
+        defs = importDef;
+      }
     }
 
     const docs = new SnapshotFragmentMap(this.languageServiceManager);
     docs.set(tsDoc.filePath, { fragment: mainFragment, snapshot: tsDoc });
 
     const result = await Promise.all(
-      defs.definitions.map(async (def) => {
+      defs.definitions!.map(async (def) => {
         const { fragment, snapshot } = await docs.retrieve(def.fileName);
 
         if (isNoTextSpanInGeneratedCode(snapshot.getFullText(), def.textSpan)) {
           const fileName = isVirtualFilePath(def.fileName) ? def.fileName.substr(0, def.fileName.length - 3) : def.fileName;
+          const textSpan = isVirtualAstroFilePath(tsFilePath) ? { start: 0, length: 0 } : def.textSpan;
           return LocationLink.create(
             pathToUrl(fileName),
-            convertToLocationRange(fragment, def.textSpan),
-            convertToLocationRange(fragment, def.textSpan),
-            convertToLocationRange(mainFragment, defs.textSpan)
+            convertToLocationRange(fragment, textSpan),
+            convertToLocationRange(fragment, textSpan),
+            convertToLocationRange(mainFragment, defs!.textSpan)
           );
         }
       })
@@ -99,6 +129,10 @@ export class TypeScriptPlugin implements CompletionsProvider {
     }
   }
 
+  async getSignatureHelp(document: Document, position: Position, context: SignatureHelpContext | undefined, cancellationToken?: CancellationToken): Promise<SignatureHelp | null> {
+    return this.signatureHelpProvider.getSignatureHelp(document, position, context, cancellationToken);
+  }
+
   /**
    *
    * @internal
@@ -109,5 +143,47 @@ export class TypeScriptPlugin implements CompletionsProvider {
 
   private isInsideFrontmatter(document: Document, position: Position) {
     return isInsideFrontmatter(document.getText(), document.offsetAt(position));
+  }
+
+  private goToDefinitionFoundOnlyAlias(tsFileName: string, defs: readonly ts.DefinitionInfo[]) {
+    return !!(defs.length === 1 && defs[0].kind === 'alias' && defs[0].fileName === tsFileName);
+  }
+
+  private getGoToDefinitionRefsForImportSpecifier(tsFilePath: string, offset: number, lang: ts.LanguageService): ts.DefinitionInfoAndBoundSpan | undefined {
+    const program = lang.getProgram();
+    const sourceFile = program?.getSourceFile(tsFilePath);
+    if (sourceFile) {
+      let node = (ts as BetterTS).getTouchingPropertyName(sourceFile, offset);
+      if (node && node.kind === SyntaxKind.Identifier) {
+        if (node.parent.kind === SyntaxKind.ImportClause) {
+          let decl = node.parent.parent as ImportDeclaration;
+          let spec = ts.isStringLiteral(decl.moduleSpecifier) && decl.moduleSpecifier.text;
+          if (spec) {
+            let fileName = pathJoin(pathDirname(tsFilePath), spec);
+            let start = node.pos + 1;
+            let def: ts.DefinitionInfoAndBoundSpan = {
+              definitions: [
+                {
+                  kind: 'alias',
+                  fileName,
+                  name: '',
+                  containerKind: '',
+                  containerName: '',
+                  textSpan: {
+                    start: 0,
+                    length: 0,
+                  },
+                } as ts.DefinitionInfo,
+              ],
+              textSpan: {
+                start,
+                length: node.end - start,
+              },
+            };
+            return def;
+          }
+        }
+      }
+    }
   }
 }