+++ tools/vscode/CHANGELOG.md
@@ -1,5 +1,16 @@
 # astro-vscode
 
+## 0.6.1
+
+- Makes the v0.6.0 features actually work 😅
+
+## 0.6.0
+
+- Fixes bug with signature help not appearing in the component script section.
+- Adds completion suggestions for `Astro.*` APIs in the component script.
+- Adds support for Hover based hints in the component script section.
+- Fixes bug with Go to Definition (cmd + click) for Components.
+
 ## 0.5.0
 
 - Bug fixes, dependency updates